[MODEL] Problem is now reset after its use in an algorithm.
This is done by creating an entirely new Problem, that implements a pseudo deep copy: all objects are copied, but all information about the room's filling and the associations between Seats and Persons are not.
Only the structure of the Problem is copied, creating a blank Problem to feed to an algorithm.